Put hardware devices in a seperate subdirectory for better code structure.
