+++ src/hardware/devices/sACNDMXDevice.h
@@ -1,25 +1,33 @@
-#ifndef DMX512_SERIAL_DEVICE_H
-#define DMX512_SERIAL_DEVICE_H
+#ifndef S_ACN_DMX_DEVICE_H
+#define S_ACN_DMX_DEVICE_H
 
 #include <SFML/System.hpp>
+#include <SFML/Network.hpp>
 #include <stdint.h>
-#include "hardwareOutputDevice.h"
+#include "hardware/hardwareOutputDevice.h"
 
-//The DMX512SerialDevice can talk to Open DMX USB hardware, and just about any hardware which is just an serial port connected to a line driver.
-class SerialPort;
-class DMX512SerialDevice : public HardwareOutputDevice
+//The AcnDMXDevice talks the ACN E1.31 protocol. Which is an UDP protocol for sending DMX messages trough IP networks.
+class StreamingAcnDMXDevice : public HardwareOutputDevice
 {
 private:
-    SerialPort* port;
+    static constexpr int acn_port = 5568;
+
     sf::Thread update_thread;
+    sf::UdpSocket socket;
     
     bool run_thread;
     int channel_count;
+    uint8_t channel_data[512];
+    
     int resend_delay;
-    uint8_t data_stream[1+512];
+    bool multicast;
+    
+    int universe;
+    uint8_t uuid[16];
+    uint8_t source_name[64];
 public:
-    DMX512SerialDevice();
-    virtual ~DMX512SerialDevice();
+    StreamingAcnDMXDevice();
+    virtual ~StreamingAcnDMXDevice();
     
     //Configure the device.
     // Parameter: port: name of the serial port to connect to.
@@ -35,4 +43,4 @@ private:
     void updateLoop();
 };
 
-#endif//DMX512_SERIAL_DEVICE_H
+#endif//S_ACN_DMX_DEVICE_H
