+++ src/hardware/devices/uDMXDevice.cpp
