+++ src/hardware/devices/virtualOutputDevice.cpp
